distribute inline keyboard helpers
